# openapi.yaml
openapi: 3.1.0
info:
  title: Regret Vault API
  description: This is my first attempt at API documentation. It allows users to register, login, and manage their regrets. Users can create regrets (public or private), view all public regrets, edit, delete their regrets. They can view their own regrets, or view a specific request by id.
  version: 1.0.0
servers:
  - url: https://regret-vault-api.onrender.com
    description: Live production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /register:
    post:
      summary: Register a new user
      description: This creates a new user in the database, this is important as you need to be registered to access other routes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                  minLength: 3
                  example: turbash
                email:
                  type: string
                  format: email
                  example: turbash@example.com
                password:
                  type: string
                  minLength: 8
                  example: turbashking
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Registration Successful
                  token:
                    type: string
                    example: jwt_token_here
        '400':
          description: User already exists
          content:
            text/html:
              schema:
                type: string
                example: User already exists
        '500':
          description: Internal Server Error
          content:
            text/html:
              schema:
                type: string
                example: Internal Server Error

  /login:
    post:
      summary: Login a user
      description: This logs in an existing user and returns a JWT token to access protected routes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: The email of the user
                  example: turbash@example.com
                password:
                  type: string
                  description: The password of the user
                  minLength: 8
                  example: turbashking
      responses:
        '200':
          description: User logged in successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Login Successful
                  token:
                    type: string
                    example: jwt_token_here
        '404':
          description: User not found
          content:
            text/html:
              schema:
                type: string
                example: User not found
        '401':
          description: Invalid email or password
          content:
            text/html:
              schema:
                type: string
                example: Invalid Credentials
  /regrets:
    post:
      summary: Create a new regret
      description: This creates a new regret entry for the authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - message
              properties:
                title:
                  type: string
                  description: The title of the regret
                  example: "My biggest regret"
                message:
                  type: string
                  description: The message describing the regret
                  example: "I wish I had studied harder."
                regretLevel:
                  type: number
                  description: The level of regret (1-10)
                  minimum: 1
                  maximum: 10
                  example: 5
                isPublic:
                  type: boolean
                  description: Whether the regret is public or private
                  example: false
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Regret created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Regret created successfully
                  data:
                    type: object
                    properties:
                      user:
                        type: string
                        example: user_id_here
                      title:
                        type: string
                        example: "My biggest regret"
                      message:
                        type: string
                        example: "I wish I had studied harder."
                      regretLevel:
                        type: number
                        example: 5
                      isPublic:
                        type: boolean
                        example: false
        '400':
          description: Invalid Token
          content:
            text/html:
              schema:
                type: string
                example: Invalid Token
        '404':
          description: User not found
          content:
            text/html:
              schema:
                type: string
                example: User not found
        '401':
          description: Unauthorized
          content:
            text/html:
              schema:
                type: string
                example: Unauthorized, Access Denied
        '500':
          description: Internal Server Error
          content:
            text/html:
              schema:
                type: string
                example: Internal Server Error
    get:
      summary: Get all public regrets
      description: This returns all the public regrets from the database in descending order of creation date i.e. most recent regrets first.
      responses:
        '200':
          description: A list of public regrets
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: number
                    example: 1
                  regrets:
                    type: array
                    items:
                      type: object
                      properties:
                        user:
                          type: string
                          example: user_id_here
                        title:
                          type: string
                          example: "My biggest regret"
                        message:
                          type: string
                          example: "I wish I had studied harder."
                        regretLevel:
                          type: number
                          example: 5
                        isPublic:
                          type: boolean
                          example: false
                        createdAt:
                          type: string
                          format: date-time
                          example: "2023-01-01T00:00:00Z"
        '404':
          description: No public regrets found
          content:
            text/html:
              schema:
                type: string
                example: No public regrets found
        '500':
          description: Internal Server Error
          content:
            text/html:
              schema:
                type: string
                example: Internal Server Error
  /my-regrets:
    get:
      summary: Get all regrets of the current user
      description: This returns all the regrets posted by the current logged in user in descending order of creation date i.e. most recent regrets first.
      security:
        - bearerAuth: []
      responses: 
        '200':
          description: A list of user's regrets public as well as private
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: number
                    example: 1
                  regrets:
                    type: array
                    items:
                      type: object
                      properties:
                        user:
                          type: string
                          example: user_id_here
                        title:
                          type: string
                          example: "My biggest regret"
                        message:
                          type: string 
                          example: "I wish I had studied harder."
                        regretLevel:
                          type: number
                          example: 5
                        isPublic:
                          type: boolean
                          example: false
                        createdAt:
                          type: string
                          format: date-time
                          example: "2023-01-01T00:00:00Z"
        '400':
          description: Invalid Token
          content:
            text/html:
              schema:
                type: string
                example: Invalid Token
        '404':
          description: No regrets found for this user
          content:
            text/html:
              schema:
                type: string
                example: No regrets found for this user
        '401':
          description: Unauthorized
          content:
            text/html:
              schema:
                type: string
                example: Unauthorized, Access Denied
        '500':
          description: Internal Server Error
          content:
            text/html:
              schema:
                type: string
                example: Internal Server Error
  /regrets/{id}:
    get:
      summary: Get a specific regret by id
      description: This returns a specific regret by its ID. If the regret is private, only the user who posted it can access it. If its public, anyone can access it.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: regret_id_here
          description: The ID of the regret to retrieve.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Regret retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: string
                    example: user_id_here
                  title:
                    type: string
                    example: "My biggest regret"
                  message:
                    type: string
                    example: "I wish I had studied harder."
                  regretLevel:
                    type: number
                    example: 5
                  isPublic:
                    type: boolean
                    example: false
                  createdAt:
                    type: string
                    format: date-time
                    example: "2023-01-01T00:00:00Z"
        '401':
          description: Unauthorized, Access Denied
          content:
            text/html:
              schema:
                type: string
                example: Unauthorized, this regret is private, if this is your regret please login
        '403':
          description: Forbidden, You do not have access to this regret
          content:
            text/html:
              schema:
                type: string
                example: Forbidden, You do not have access to this regret
        '404':
          description: Regret not found
          content:
            text/html:
              schema:
                type: string
                example: Regret not found
        '500':
          description: Internal Server Error
          content:
            text/html:
              schema:
                type: string
                example: Internal Server Error
    put:
      summary: Update a specific regret by id
      description: This updates a specific regret by taking id as path parameter. Only the user who created the regret can update it.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: regret_id_here
          description: The ID of the regret to update.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: The title of the regret
                  example: "My biggest regret"
                message:
                  type: string
                  description: The message describing the regret
                  example: "I wish I had studied harder."
                regretLevel:
                  type: number
                  description: The level of regret (1-10)
                  minimum: 1
                  maximum: 10
                  example: 5
                isPublic:
                  type: boolean
                  description: Whether the regret is public or private
                  example: false
      responses:
        '200':
          description: Regret updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Regret updated successfully
                  data:
                    type: object 
                    properties:
                      user:
                        type: string
                        example: user_id_here
                      title:
                        type: string
                        example: "My biggest regret"
                      message:
                        type: string
                        example: "I wish I had studied harder."
                      regretLevel:
                        type: number
                        example: 5
                      isPublic:
                        type: boolean
                        example: false
        '400':
          description: Invalid Token
          content:
            text/html:
              schema:
                type: string
                example: Invalid Token
        '404':
          description: Regret not found
          content:
            text/html:
              schema:
                type: string
                example: Regret not found
        '401':
          description: Unauthorized
          content:
            text/html:
              schema:
                type: string
                example: Unauthorized, Access Denied
        '403':
          description: Forbidden
          content:
            text/html:
              schema:
                type: string
                example: Forbidden, you do not have access to edit this regret
        '500':
          description: Internal Server Error
          content:
            text/html:
              schema:
                type: string
                example: Internal Server Error
    delete:
      summary: Delete a specific regret by id
      description: This deletes a specific regret by its ID. Only the user who posted the regret can delete it. Requires user to be logged in.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: regret_id_here
          description: The ID of the regret to update.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Regret deleted successfully
          content: 
            text/html:
              schema:
                type: string
                example: Regret deleted successfully
        '400':
          description: Invalid token
          content:
            text/html:
              schema:
                type: string
                example: Invalid Token
        '401':
          description: Unauthorized
          content:
            text/html:
              schema:
                type: string
                example: Unauthorized, Access Denied
        '403':
          description: Forbidden
          content:
            text/html:
              schema:
                type: string
                example: Forbidden, You do not have access to this regret
        '404':
          description: Regret not found
          content:
            text/html:
              schema:
                type: string
                example: Regret not found
        '500':
          description: Internal Server Error
          content:
            text/html:
              schema:
                type: string
                example: Internal Server Error
  /regrets/search:
    get:
      summary: Search regrets by title or message
      description: This allows users to search for specific public regrets by their title or message.
      parameters:
        - in: query
          name: query
          required: true
          schema:
            type: string
            example: "studied."
          description: The search query to filter regrets by title or message.
      responses:
        '200':
          description: All public regrets which match the search query
          content: 
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: number
                    example: 2
                  regrets:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: regret_id_here
                        title:
                          type: string
                          example: "I regret studying for exams."
                        message:
                          type: string
                          example: "I wish I had spent more time with my friends."
                        createdAt:
                          type: string
                          format: date-time
                          example: "2023-10-01T12:00:00Z"
                        updatedAt:
                          type: string
                          format: date-time
                          example: "2023-10-01T12:00:00Z"
        '400':
          description: Search query is required
          content:
            text/html:
              schema:
                type: string
                example: Search query is required
        '404':
          description: No regrets found matching the search query
          content:
            text/html:
              schema:
                type: string
                example: No regrets found matching the search query
        '500':
          description: Internal Server Error
          content:
            text/html:
              schema:
                type: string      
                example: Internal Server Error
  /my-regrets/search:
    get:
      summary: Search your regrets by title or message
      description: This allows users to search for their own regrets by title or message. You need to be authenticated.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: query
          required: true
          schema:
            type: string
            example: "studied."
          description: The search query to filter regrets by title or message.
      responses:
        '200':
          description: All your regrets which match the search query
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: number
                    example: 2
                  regrets:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: regret_id_here
                        title:
                          type: string
                          example: "I regret studying for exams."
                        message:
                          type: string
                          example: "I wish I had spent more time with my friends."
                        createdAt:
                          type: string
                          format: date-time
                          example: "2023-10-01T12:00:00Z"
                        updatedAt:
                          type: string
                          format: date-time
                          example: "2023-10-01T12:00:00Z"
        '400':
          description: Search query is required
          content:
            text/html:
              schema:
                type: string
                example: Search query is required
        '401':
          description: Unauthorized
          content:
            text/html:
              schema:
                type: string
                example: Unauthorized, Access Denied
        '404':
          description: No regrets found matching the search query
          content:
            text/html:
              schema:
                type: string
                example: No regrets found matching the search query
        '500':
          description: Internal Server Error
          content:
            text/html:
              schema:
                type: string
                example: Internal Server Error